+++ SplitWiseImpl/src/com/factory/interfaces/IStrategyListFactory.java
@@ -1,6 +1,8 @@
-package com.interfaces;
+package com.factory.interfaces;
 
 import java.util.List;
+
+import com.interfaces.IStrategyList;
 
 public interface IStrategyListFactory {
 	IStrategyList getStrategyList(List<String> strategiesPresent);
