--- SplitWiseImpl/src/com/interfaces/IDataStoreFactory.java
@@ -1,7 +1,0 @@
-package com.interfaces;
-
-public interface IDataStoreFactory {
-	IUserDataStore getUserDataStore();
-	
-	IUserPairOwedAmountDataStore getUserPairOwedAmountDataStore();
-}
