+++ SplitWiseImpl/src/com/ioParser/interfaces/IInputReader.java
@@ -1,6 +1,8 @@
-package com.interfaces;
+package com.ioParser.interfaces;
 
 import java.util.List;
+
+import com.strategies.interfaces.ISplitInput;
 
 public interface IInputReader {
 	ISplitInput parseInput(List<String> input);
