+++ SplitWiseImpl/src/com/strategies/equalDivisionStrategy/EqualStrategyInutGenerator.java
@@ -1,10 +1,10 @@
-package com.equalDivisionStrategy;
+package com.strategies.equalDivisionStrategy;
 
 import java.util.List;
 
-import com.interfaces.ISplitInput;
-import com.splitObjects.SplitInput;
-import com.splitObjects.StrategyInputGenerator;
+import com.strategies.interfaces.ISplitInput;
+import com.strategies.splitObjects.SplitInput;
+import com.strategies.splitObjects.StrategyInputGenerator;
 
 public class EqualStrategyInutGenerator extends StrategyInputGenerator {
