+++ SplitWiseImpl/src/com/strategies/percentDivisionStrategy/PercentDivisionSplitInput.java
@@ -1,10 +1,10 @@
-package com.percentDivisionStrategy;
+package com.strategies.percentDivisionStrategy;
 
 import java.util.HashMap;
 import java.util.List;
 
 import com.enums.SplitStrategyType;
-import com.splitObjects.SplitInput;
+import com.strategies.splitObjects.SplitInput;
 
 public class PercentDivisionSplitInput extends SplitInput {
 	private HashMap<String, Double> percentOwedMap;
